fix error at ex008 (change int to float)

+++ workouts/ex008.py
@@ -1,5 +1,5 @@
 # Escreva um programa que leia um valor em metros e faça a conversão para centímetros e milímetros # noqa: E501
-meter = int(input('Digite a metragem que deseja calcular: '))
+meter = float(input('Digite a metragem que deseja calcular: '))
 cent = meter * 100
 mili = meter * 1000
